refactor(app): organize sources for `timeline` page
