+++ app/src/index.tsx
@@ -1,7 +1,7 @@
 import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import Timeline from "./Timeline";
+import Timeline from "./timeline/Timeline";
 import "./index.css";
 
 const root = document.getElementById("root");
